change GraphPoint.y to Int

+++ cps-compose/app/src/main/java/com/demich/cps/profiles/rating_graph/GraphPoint.kt
@@ -5,8 +5,8 @@ import kotlin.time.Instant
 
 internal data class GraphPoint(
     val x: Instant,
-    val y: Long
+    val y: Int
 )
 
 internal fun RatingChange.toGraphPoint() =
-    GraphPoint(x = date, y = rating.toLong())
+    GraphPoint(x = date, y = rating)
